pindahkan supplier ke component
