+++ web/src/base-components/Form/FormSupplier/index.ts
@@ -1,0 +1,3 @@
+import FormSupplier from './FormSupplier.vue'
+
+export default FormSupplier
